Split old LegacyErpRedeemScriptBuilder into NonStandardErp builders related to activations
